requirements.txt
mcp
caldav
icalendar
python-dotenv
